// src/app/models/job.ts
import { Time } from '@angular/common';

export class Job {
    _id: string;
    title: string;
    applylink: string;
    jd: string;
    companyname: string;
    location: string;
    experience: string;
    salary: string;
    type: string;
    skills: string;
    startdate: string;
    enddate: string;
    created: string;
    source: string;
    timestamp: string;
    __v: string;
}